combined validateH5.rs into readH5.rs
