+++ rust/Cargo.toml
@@ -113,9 +113,6 @@ name="cerno"
 path="src/cerno.rs"
 
 [[bin]]
-name="validateH5"
-path="src/validateH5.rs"
-
-[[bin]]
 name="readH5"
 path="src/readH5.rs"
+
